+++ Kainote/AudioPlayerDSound.h
@@ -1,43 +1,25 @@
-// Copyright (c) 2008, Niels Martin Hansen
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
-//
-//   * Redistributions of source code must retain the above copyright notice,
-//     this list of conditions and the following disclaimer.
-//   * Redistributions in binary form must reproduce the above copyright notice,
-//     this list of conditions and the following disclaimer in the documentation
-//     and/or other materials provided with the distribution.
-//   * Neither the name of the Aegisub Group nor the names of its contributors
-//     may be used to endorse or promote products derived from this software
-//     without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
-// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-// POSSIBILITY OF SUCH DAMAGE.
-//
-// -----------------------------------------------------------------------------
-//
-// AEGISUB
-//
-// Website: http://www.aegisub.net/
-// Contact: mailto:jiifurusu@gmail.com
-//
+//  Copyright (c) 2016-2022, Marcin Drob
+
+//  Kainote is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+
+//  Kainote is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+
+//  You should have received a copy of the GNU General Public License
+//  along with Kainote.  If not, see <http://www.gnu.org/licenses/>.
+
+//this code piervously was taken from Aegisub 2 it's rewritten by me almost all.
 
 #pragma once
 
 #include <wx/wxprec.h>
-#include <stdint.h>
-#include <mmsystem.h>
+//#include <stdint.h>
+//#include <mmsystem.h>
 #include <dsound.h>
 
 #include "Provider.h"
@@ -62,16 +44,16 @@ public:
 	void CloseStream();
 	void SetProvider(Provider *_provider);
 
-	void Play(int64_t start, int64_t count);
+	void Play(long long start, long long count);
 	void Stop(bool timerToo = true);
 	bool IsPlaying();
 
-	int64_t GetStartPosition();
-	int64_t GetEndPosition();
-	int64_t GetCurrentPosition();
+	long long GetStartPosition();
+	long long GetEndPosition();
+	long long GetCurrentPosition();
 	int GetCurPositionMS();
-	void SetEndPosition(int64_t pos);
-	void SetCurrentPosition(int64_t pos);
+	void SetEndPosition(long long pos);
+	void SetCurrentPosition(long long pos);
 
 	void SetVolume(double vol);
 	double GetVolume();
